+++ database/factories/WikiNotificationSentRecordFactory.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Database\Factories;
+
+use App\Notifications\EmptyWikiNotification;
+use App\WikiNotificationSentRecord;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class WikiNotificationSentRecordFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = WikiNotificationSentRecord::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [];
+    }
+}